pyproject.toml
[project]
name = "prompt-management-system"
version = "0.1.0"
description = "A prompt management system with MCP server support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "google-generativeai>=0.3.0",
    "openai>=1.6.0",
    "aiofiles>=23.2.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "sse-starlette>=1.8.0",
    "pydantic-settings>=0.2.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=3.2.0,<4.0.0",
    "python-jose[cryptography]>=3.3.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.28.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.8",
]

[tool.hatch.build.targets.wheel]
force-include = { backend = "prompt_management_system" }
